Split up mocking files
